Cập nhật

+++ Qly_NhaHang/Qly_NhaHang/frmDetailRecipe.cs
@@ -19,7 +19,6 @@ namespace Qly_NhaHang
 
         public frmDetailRecipe(string nameIngredient, string unitIngredient, int ingredientCount, int idFood)
         {
-            
             InitializeComponent();
             lblNameIngre.Text = nameIngredient;
             lblUnit.Text = unitIngredient;
@@ -38,7 +37,7 @@ namespace Qly_NhaHang
                 {
                     return ingredient.id_Ingredient;
                 }
-                return -1; // Trả về một giá trị không hợp lệ nếu không tìm thấy món ăn
+                return -1; 
             }
         }
 
@@ -49,59 +48,57 @@ namespace Qly_NhaHang
                 int idIngredient = GetIngredientIdByName(lblNameIngre.Text);
                 string idFoodText = lblID.Text;
                int idFoodValue = int.Parse(idFoodText);
-
-
-                Recipe existingRecipe = dbContext.Recipes
-                    .FirstOrDefault(bi => bi.id_Food == idFoodValue && bi.id_Ingredient == idIngredient);
-
-                if (existingRecipe == null)
+                if (idFoodValue == 0)
                 {
-                    // Tạo mới một BillInfo nếu không tồn tại
-                    if (ingredientCount == 0 && quantity == 0)
+                    XtraMessageBox.Show("Vui lòng chọn món ăn để thiết lập công thức", "Lỗi", MessageBoxButtons.OKCancel, MessageBoxIcon.Error);
+                }
+                else 
+                {
+                    Recipe existingRecipe = dbContext.Recipes
+                            .FirstOrDefault(bi => bi.id_Food == idFoodValue && bi.id_Ingredient == idIngredient);
+                    if (existingRecipe == null)
                     {
-                        XtraMessageBox.Show("Vui lòng nhập số lượng món.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                        return;
-                    }
-                    else
-                    {
-                        if (ingredientCount == 0)
+                        // Tạo mới một BillInfo nếu không tồn tại
+                        if (ingredientCount == 0 && quantity == 0)
                         {
-                            Recipe newRecipe = new Recipe
+                            XtraMessageBox.Show("Vui lòng nhập số lượng món.", "Thông báo", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                            return;
+                        }
+                        else
+                        {
+                            if (ingredientCount == 0)
                             {
-                                id_Food = idFoodValue,
-                                id_Ingredient = idIngredient,
-                                count_Ingredient = quantity,
-                            };
-                            dbContext.Recipes.Add(newRecipe);
-                            dbContext.SaveChanges();
+                                Recipe newRecipe = new Recipe
+                                {
+                                    id_Food = idFoodValue,
+                                    id_Ingredient = idIngredient,
+                                    count_Ingredient = quantity,
+                                };
+                                dbContext.Recipes.Add(newRecipe);
+                                dbContext.SaveChanges();
+                            }
                         }
                     }
-                }
-                else
-                {
-                    if (nmrQuantity.Value == 0)
-                    {
-                        // Xóa BillInfo nếu foodCount bằng 0
-                        dbContext.Recipes.Remove(existingRecipe);
-                    }
                     else
                     {
-                        // Cập nhật count_Food nếu foodCount khác 0
-                        existingRecipe.count_Ingredient = quantity;
+                        if (nmrQuantity.Value == 0)
+                        {
+                            dbContext.Recipes.Remove(existingRecipe);
+                        }
+                        else
+                        {
+                            existingRecipe.count_Ingredient = quantity;
+                        }
+                    }
+                    dbContext.SaveChanges();
+                    this.Close();
+                    if (Application.OpenForms["frmRecipe"] is frmRecipe recipeForm)
+                    {
+                        recipeForm.LoadDataRecipe(); 
+                        recipeForm.GetIngredientCount(ingredientCount);
+                        recipeForm.LoadIngredientFLPNTest();
                     }
                 }
-
-                dbContext.SaveChanges();
-
-                this.Close();
-                if (Application.OpenForms["frmRecipe"] is frmRecipe recipeForm)
-                {
-                    recipeForm.LoadDataRecipe(); // Giả sử tên phương thức là LoadFoodData()
-
-                    recipeForm.GetIngredientCount(ingredientCount);
-                    recipeForm.LoadIngredientFLPNTest();
-                }
-
             }
         }
